requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2020.11.8
clipboard==0.0.4
cloudinary==1.23.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-registration==3.1.1
django-tinymce==3.1.0
django-utils-six==2.0
djangorestframework==3.12.2
gunicorn==20.0.4
importlib-metadata==2.0.0
Pillow==8.0.1
psycopg2==2.8.6
pyperclip==1.8.1
python-decouple==3.1
pytz==2020.4
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
zipp==3.4.0
